WIP ...  Improvement of the favicon configuration
a. Reduce default LIMIT_TOTAL_BYTES to 50MB (before it was 1GB)

b. The RESOLVER_MAP has been moved to the configuration
   [favicons.proxy.resolver_map] in the favicons.toml file for three reasons:

    1. by default only the resolver from search.favicon_resolver is available to
       the user.

    2. more resolvers can be activated in the favicons.toml

    3. Additional resolvers can be integrated (resolvers that are not implemented
       in the SearXNG core).

With the positions a. and b.1, a simple and less resource-hungry default
configuration should be made possible, where a maintainer only have to set
search.favicon_resolver in the settings.

Signed-off-by: Markus <markus@venom.fritz.box>